requirements.txt
robotframework==5.0.1
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.3
robotframework-stacktrace==0.4.1
lxml==4.9.0
glob2==0.7